+++ examples/src/bin/etm_gpio.rs
@@ -15,6 +15,7 @@ use esp_hal::{
         etm::{GpioEtmChannels, GpioEtmInputConfig, GpioEtmOutputConfig},
         Io,
         Level,
+        Output,
         Pull,
     },
     prelude::*,
@@ -25,7 +26,7 @@ fn main() -> ! {
     let peripherals = esp_hal::init(esp_hal::Config::default());
 
     let io = Io::new(peripherals.GPIO, peripherals.IO_MUX);
-    let mut led = io.pins.gpio1;
+    let mut led = Output::new(io.pins.gpio1, Level::Low);
     let button = io.pins.gpio9;
 
     led.set_high();